started backend bin and moved thihngs around
