+++ backend/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "smithe_backend"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+diesel = { version = "1.4", features = ["postgres"] }
+rocket = { version = "0.5.0-rc.2", features = ["json"] }
+smithe_database = { path = "../database" }
+startgg = { path = "../startgg" }
