+++ database/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "smithe_database"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+diesel = { version = "1.4", features = ["postgres"] }
+serde = "1"
+startgg = { path = "../startgg" }
