+++ database/src/db_models/game.rs
@@ -1,7 +1,8 @@
 #![allow(clippy::extra_unused_lifetimes)]
 // ^^^ this is needed because Insertable introduces a lifetime we don't use
 // — an auto fix for this exists only in Diesel v2.
-use crate::{schema::player_games, startgg::SSBU_CHARACTERS};
+use crate::schema::player_games;
+use startgg::SSBU_CHARACTERS;
 
 #[derive(Debug, Insertable, Queryable, QueryableByName)]
 #[table_name = "player_games"]
