+++ src/smithe/event.rs
@@ -1,4 +1,4 @@
-use crate::queries::{
+use startgg::queries::{
     event_getter::get_phase_id_from_event_slug, seeding_geeter::make_seeding_getter_query,
 };
 use anyhow::Result;
