+++ startgg/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "startgg"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+gql_client = "1"
+serde = "1"
+tokio = { version = "1.1", features = ["full"] }
